backend/user-management/match-history-service/requirements.txt
asgiref==3.8.1
Django>=5.1.4
djangorestframework>=3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
django-extensions==3.2.3
Werkzeug>=3.0.6
pyOpenSSL==23.2.0
Pillow==11.0.0
setuptools==75.3.0
channels_redis==4.0.0
redis==5.0.0
requests>=2.0.0
